+++ lib/flutter_map_marker_popup.dart
@@ -1,6 +1,7 @@
 library flutter_map_marker_popup;
 
 export 'src/marker_center_animation.dart';
+export 'src/marker_tap_behavior.dart';
 export 'src/popup_animation.dart';
 export 'src/popup_controller.dart';
 export 'src/popup_marker_layer_options.dart';