+++ lib/src/popup_container/marker_with_key.dart
@@ -8,4 +8,15 @@ class MarkerWithKey {
   final Marker marker;
 
   MarkerWithKey(this.marker) : key = GlobalKey();
+
+  static MarkerWithKey wrap(Marker marker) => MarkerWithKey(marker);
+
+  static Marker unwrap(MarkerWithKey markerWithKey) => markerWithKey.marker;
+
+  @override
+  bool operator ==(Object other) =>
+      other is MarkerWithKey && marker == other.marker;
+
+  @override
+  int get hashCode => marker.hashCode;
 }