+++ lib/src/popup_marker_layer_widget.dart
@@ -4,6 +4,7 @@ import 'package:flutter_map_marker_popup/src/popup_layer.dart';
 
 import '../flutter_map_marker_popup.dart';
 import 'marker_layer.dart';
+import 'popup_controller_impl.dart';
 
 class PopupMarkerLayerWidget extends StatelessWidget {
   final PopupMarkerLayerOptions options;
@@ -14,13 +15,18 @@ class PopupMarkerLayerWidget extends StatelessWidget {
   Widget build(BuildContext context) {
     final mapState = flutter_map.MapState.maybeOf(context)!;
     return Stack(children: [
-      MarkerLayer(options, mapState, mapState.onMoved, options.popupController),
+      MarkerLayer(
+        options,
+        mapState,
+        mapState.onMoved,
+        options.popupController as PopupControllerImpl,
+      ),
       PopupLayer(
         mapState: mapState,
         stream: mapState.onMoved,
         popupSnap: options.popupSnap,
         popupBuilder: options.popupBuilder,
-        popupController: options.popupController,
+        popupController: options.popupController as PopupControllerImpl,
         popupAnimation: options.popupAnimation,
         markerRotate: options.rotate ?? false,
       ),